重构项目

+++ clouddo-admin/src/main/java/com/bootdo/clouddoadmin/service/impl/RoleServiceImpl.java
@@ -6,6 +6,7 @@ import com.bootdo.clouddoadmin.dao.UserDao;
 import com.bootdo.clouddoadmin.dao.UserRoleDao;
 import com.bootdo.clouddoadmin.domain.RoleDO;
 import com.bootdo.clouddoadmin.domain.RoleMenuDO;
+import com.bootdo.clouddoadmin.dto.UserRoleDTO;
 import com.bootdo.clouddoadmin.service.RoleService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.CacheEvict;
@@ -13,10 +14,7 @@ import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Objects;
+import java.util.*;
 
 
 @Service
@@ -35,29 +33,40 @@ public class RoleServiceImpl implements RoleService {
     @Autowired
     UserRoleDao userRoleMapper;
 
-    @Cacheable(value = DEMO_CACHE_NAME, key = ROLE_ALL_KEY)
     @Override
-    public List<RoleDO> list() {
-        List<RoleDO> roles = roleMapper.list(new HashMap<>(16));
+    public List<RoleDO> list(Map<String, Object> map) {
+        List<RoleDO> roles = roleMapper.list(map);
         return roles;
+    }
+
+    @Override
+    public int count(Map<String, Object> map) {
+        return roleMapper.count(map);
     }
 
 
     @Override
-    public List<RoleDO> list(Long userId) {
+    public List<UserRoleDTO> list(Long userId) {
         List<Long> rolesIds = userRoleMapper.listRoleId(userId);
         List<RoleDO> roles = roleMapper.list(new HashMap<>(16));
+        List<UserRoleDTO> roleDTOS = new ArrayList<>();
         for (RoleDO roleDO : roles) {
-            roleDO.setRoleSign("false");
+            UserRoleDTO userRoleDTO = new UserRoleDTO();
+            userRoleDTO.setId(roleDO.getRoleId());
+            userRoleDTO.setName(roleDO.getRoleName());
+            userRoleDTO.setChecked(false);
             for (Long roleId : rolesIds) {
                 if (Objects.equals(roleDO.getRoleId(), roleId)) {
-                    roleDO.setRoleSign("true");
+                    // roleDO.setRoleSign("true");
+                    userRoleDTO.setChecked(true);
                     break;
                 }
             }
+            roleDTOS.add(userRoleDTO);
         }
-        return roles;
+        return roleDTOS;
     }
+
     @CacheEvict(value = DEMO_CACHE_NAME, key = ROLE_ALL_KEY)
     @Transactional
     @Override
@@ -79,7 +88,6 @@ public class RoleServiceImpl implements RoleService {
         return count;
     }
 
-    @CacheEvict(value = DEMO_CACHE_NAME, key = ROLE_ALL_KEY)
     @Transactional
     @Override
     public int remove(Long id) {
@@ -94,22 +102,23 @@ public class RoleServiceImpl implements RoleService {
         return roleDO;
     }
 
-    @CacheEvict(value = DEMO_CACHE_NAME, key = ROLE_ALL_KEY)
     @Override
     public int update(RoleDO role) {
         int r = roleMapper.update(role);
         List<Long> menuIds = role.getMenuIds();
-        Long roleId = role.getRoleId();
-        roleMenuMapper.removeByRoleId(roleId);
-        List<RoleMenuDO> rms = new ArrayList<>();
-        for (Long menuId : menuIds) {
-            RoleMenuDO rmDo = new RoleMenuDO();
-            rmDo.setRoleId(roleId);
-            rmDo.setMenuId(menuId);
-            rms.add(rmDo);
-        }
-        if (rms.size() > 0) {
-            roleMenuMapper.batchSave(rms);
+        if(null!=menuIds){
+            Long roleId = role.getRoleId();
+            roleMenuMapper.removeByRoleId(roleId);
+            List<RoleMenuDO> rms = new ArrayList<>();
+            for (Long menuId : menuIds) {
+                RoleMenuDO rmDo = new RoleMenuDO();
+                rmDo.setRoleId(roleId);
+                rmDo.setMenuId(menuId);
+                rms.add(rmDo);
+            }
+            if (rms.size() > 0) {
+                roleMenuMapper.batchSave(rms);
+            }
         }
         return r;
     }
@@ -118,6 +127,17 @@ public class RoleServiceImpl implements RoleService {
     public int batchremove(Long[] ids) {
         int r = roleMapper.batchRemove(ids);
         return r;
+    }
+
+    /**
+     * 获取用户的角色id
+     *
+     * @param userId
+     * @return 角色id
+     */
+    @Override
+    public List<Long> RoleIdsByUserId(Long userId) {
+        return roleMapper.roleIdsByUserId(userId);
     }
 
 }
